requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
psycopg==3.1.4
psycopg2-binary==2.9.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
